- Теперь метод loadFromFile, в классе FileBackedTaskManager, восстанавливает максимальный id - Теперь метод loadFromFile, в классе FileBackedTaskManager, не использует метод для создания задачи, а добавляет задачи напрямую в хранилище - Собственное исключение ManagerSaveException вынесено в отдельный класс - Реализован в тестах абстрактный класс AbstractTaskManagerTest
