+++ src/manager/ManagerSaveException.java
@@ -1,0 +1,8 @@
+package manager;
+
+public class ManagerSaveException extends RuntimeException {
+
+    public ManagerSaveException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
